+++ apps/website/src/components/blocks/overview-page.tsx
@@ -1,8 +1,11 @@
 import Head from 'next/head'
 import { CATEGORIES } from '../../data/blocks'
 import { CategoriesList } from './categories-list'
-import { Box, Container, Heading, Text } from '@chakra-ui/react'
+import { Box, Container, Heading, Text, HStack } from '@chakra-ui/react'
 import { BackgroundGradientRadial } from '../background-gradient-radial'
+import { ButtonLink } from '../link'
+import { FaFigma } from 'react-icons/fa'
+import { FigmaButton } from './figma-button'
 
 interface OverviewPageProps {
   componentsCountByCategory: Record<string, number>
@@ -30,17 +33,21 @@ export function OverviewPage({ componentsCountByCategory }: OverviewPageProps) {
       />
 
       <Box position="relative" zIndex="1">
-        <Box mb="12">
-          <Heading as="h2" size="2xl" textStyle="pageTitle">
-            Blocks
-          </Heading>
-          <Text fontSize="xl" mt="5" color="muted">
-            Build delightful apps and dashboards faster with pre-built
-            templates.
-            <br />
-            New blocks released every 2 weeks.
-          </Text>
-        </Box>
+        <HStack alignItems="start">
+          <Box mb="12" flex="1">
+            <Heading as="h2" size="2xl" textStyle="pageTitle">
+              Blocks
+            </Heading>
+            <Text fontSize="xl" mt="5" color="muted">
+              Build delightful apps and dashboards faster with pre-built
+              templates.
+              <br />
+              New blocks released regularly.
+            </Text>
+          </Box>
+
+          <FigmaButton />
+        </HStack>
 
         <CategoriesList
           groups={CATEGORIES}