+++ apps/website/src/utils/track-event.ts
@@ -1,0 +1,7 @@
+export const trackEvent = (category: string, meta: Record<string, any>) => {
+  if (typeof window === 'undefined') {
+    return
+  }
+  /* @ts-ignore */
+  window.pirsch?.(category, { meta })
+}