+++ packages/saas-ui-forms/src/base-field.tsx
@@ -9,9 +9,11 @@ import {
   FormErrorMessage,
 } from '@chakra-ui/react'
 
+import { splitProps } from '@saas-ui/core'
+
 import { useFormContext } from './form-context'
 
-import { BaseFieldProps, FieldProps } from './types'
+import type { BaseFieldProps } from './types'
 
 const getError = (name: string, formState: FormState<{ [x: string]: any }>) => {
   return get(formState.errors, name)
@@ -24,22 +26,42 @@ const isTouched = (
   return get(formState.touchedFields, name)
 }
 
+export const useBaseField = (props: BaseFieldProps) => {
+  const [fieldProps] = splitProps(props, ['name', 'label', 'help', 'hideLabel'])
+
+  const [controlProps] = splitProps(props, [
+    'id',
+    'isDisabled',
+    'isInvalid',
+    'isReadOnly',
+    'isRequired',
+  ])
+
+  const { formState } = useFormContext()
+
+  const error = getError(fieldProps.name, formState)
+  const touched = isTouched(fieldProps.name, formState)
+
+  return {
+    ...fieldProps,
+    controlProps,
+    error,
+    touched,
+  }
+}
+
 /**
  * The default BaseField component
  * Composes the Chakra UI FormControl component, with FormLabel, FormHelperText and FormErrorMessage.
  */
 export const BaseField: React.FC<BaseFieldProps> = (props) => {
-  const { name, label, help, hideLabel, children, ...controlProps } = props
-
-  const { formState } = useFormContext()
-
-  const error = getError(name, formState)
+  const { controlProps, label, help, hideLabel, error } = useBaseField(props)
 
   return (
     <FormControl {...controlProps} isInvalid={!!error}>
       {label && !hideLabel ? <FormLabel>{label}</FormLabel> : null}
       <Box>
-        {children}
+        {props.children}
         {help && !error?.message ? (
           <FormHelperText>{help}</FormHelperText>
         ) : null}