+++ apps/palette/CHANGELOG.md
@@ -1,5 +1,52 @@
 # palette
 
+## 1.4.31
+
+### Patch Changes
+
+- Updated dependencies [86caa6f]
+  - @saas-ui/theme-glass@0.5.5
+
+## 1.4.30
+
+### Patch Changes
+
+- @saas-ui/react@2.7.4
+- @saas-ui/theme-glass@0.5.4
+
+## 1.4.29
+
+### Patch Changes
+
+- Updated dependencies [20c7175]
+- Updated dependencies [ca911e4]
+- Updated dependencies [0a218fd]
+  - @saas-ui/react@2.7.3
+  - @saas-ui/theme-glass@0.5.4
+
+## 1.4.28
+
+### Patch Changes
+
+- Updated dependencies [d77a5b3]
+  - @saas-ui/react@2.7.2
+  - @saas-ui/theme-glass@0.5.4
+
+## 1.4.27
+
+### Patch Changes
+
+- @saas-ui/react@2.7.1
+- @saas-ui/theme-glass@0.5.4
+
+## 1.4.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/react@2.7.0
+  - @saas-ui/theme-glass@0.5.4
+
 ## 1.4.25
 
 ### Patch Changes