+++ packages/saas-ui-data-table/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @saas-ui/data-table
 
+## 9.0.4
+
+### Patch Changes
+
+- @saas-ui/react@2.7.4
+
+## 9.0.3
+
+### Patch Changes
+
+- Updated dependencies [20c7175]
+- Updated dependencies [ca911e4]
+- Updated dependencies [0a218fd]
+  - @saas-ui/core@2.4.4
+  - @saas-ui/react@2.7.3
+
+## 9.0.2
+
+### Patch Changes
+
+- Updated dependencies [d77a5b3]
+  - @saas-ui/core@2.4.3
+  - @saas-ui/react@2.7.2
+
+## 9.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/core@2.4.2
+  - @saas-ui/react@2.7.1
+
+## 9.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/react@2.7.0
+
 ## 8.0.2
 
 ### Patch Changes